playMedia.js -> playMedia.ts
