+++ src/mod/modules/GeneralModule.js
@@ -3,7 +3,7 @@ import jishoOrgLookup from '../../commands/jishoOrgLookup.ts'
 import madochanCreateWord from '../../commands/madochanCreateWord.ts'
 import text from '../../commands/text.ts'
 import randomText from '../../commands/randomText.ts'
-import playMedia from '../../commands/playMedia.js'
+import playMedia from '../../commands/playMedia.ts'
 import fn from '../../fn.ts'
 import chatters from '../../commands/chatters.ts'
 import Db from '../../Db.ts'
